# internal/server/openapi.yaml
openapi: "3.1.0"
info:
  title: ob serve API
  version: 0.1.0
  description: Local HTTP API for the OpenBindings CLI (ob serve).
servers:
  - url: ${OB_SERVER_URL}
    description: Default local server (port 20290 = 0x4F42, ASCII OB)
paths:
...
  /bindings/execute:
    post:
      operationId: executeBinding
      summary: Execute a resolved binding directly
      description: >
        Low-level binding execution endpoint. Accepts a binding source, ref, input,
        and optional context, then delegates to the appropriate binding executor.
        This is the binding-executor host capability.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source, ref]
              properties:
                source:
                  $ref: "#/components/schemas/ExecuteSource"
                ref:
                  type: string
                  description: Reference to the binding within the source (e.g., JSON Pointer for OpenAPI)
                input:
                  type: object
                  description: Operation input data
                context:
                  type: object
                  description: Binding context (credentials, tokens)
                  additionalProperties: true
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
        "400":
          description: Invalid input or execution error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /interfaces/create:
    post:
      operationId: createInterface
      summary: Create an OpenBindings interface from binding sources
      description: >
        Creates an OpenBindings interface document, optionally from one or more
        binding source artifacts. This is the interface-creator host capability.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                openbindingsVersion:
                  type: string
                  description: Target OpenBindings spec version (e.g., "0.1.0")
                sources:
                  type: array
                  items:
                    $ref: "#/components/schemas/CreateInterfaceSource"
                name:
                  type: string
                  description: Interface name (overrides value derived from sources)
                version:
                  type: string
                  description: Interface contract version
                description:
                  type: string
                  description: Interface description
      responses:
        "200":
          description: Created OpenBindings interface document
          content:
            application/json:
              schema:
                type: object
                description: A valid OpenBindings interface document
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /validate:
    post:
      operationId: validateInterface
      summary: Validate an OpenBindings interface document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [locator]
              properties:
                locator:
                  type: string
                strict:
                  type: boolean
      responses:
        "200":
          description: Validation report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationReport"
  /diff:
    post:
      operationId: compareInterfaces
      summary: Structural comparison of two OBIs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [baseline]
              properties:
                baseline:
                  type: string
                comparison:
                  type: string
                fromSources:
                  type: boolean
                onlySource:
                  type: string
      responses:
        "200":
          description: Diff report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiffReport"
  /compatibility:
    post:
      operationId: assessCompatibility
      summary: Check interface conformance between two interfaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target, candidate]
              properties:
                target:
                  type: string
                candidate:
                  type: string
      responses:
        "200":
          description: Compatibility report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompatibilityReport"
...
  /http/request:
    post:
      operationId: request
      summary: Make an HTTP request and return the response
      description: >
        Makes an HTTP request on behalf of the caller and returns the complete
        response. Useful for clients that cannot make direct HTTP requests due
        to platform constraints (browser CSP/CORS, network restrictions, etc.).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HttpRequestInput"
      responses:
        "200":
          description: HTTP response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Static session token (printed on startup or set via --token/OB_SERVE_TOKEN)
        or an OAuth2 access token obtained via the /oauth/authorize + /oauth/token flow.
  schemas:
...
    ValidationReport:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
    DiffReport:
      type: object
      properties:
        identical:
          type: boolean
        operations:
          type: array
          items:
            type: object
    CompatibilityReport:
      type: object
      properties:
        compatible:
          type: boolean
        issues:
          type: array
          items:
            type: object
    ExecutionResult:
      type: object
      properties:
        output:
          type: object
        error:
          type: object
          nullable: true
    ExecuteSource:
      type: object
      required: [format]
      properties:
        format:
          type: string
          description: Format token (e.g., openapi@3.1, usage@2.13.1)
        location:
          type: string
          description: Location of the binding source (path or URL)
        content:
          description: Binding source content directly
    CreateInterfaceSource:
      type: object
      required: [format]
      properties:
        format:
          type: string
          description: Format token for this source
        location:
          type: string
          description: Location to fetch content from (path or URL)
        name:
          type: string
          description: Key name for this binding source in the output OBI's sources
        outputLocation:
          type: string
          description: Location to use in output instead of input path
        description:
          type: string
          description: Human-readable description for this binding source
        embed:
          type: boolean
          description: If true, embed content inline in the output OBI
    HttpRequestInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: The URL to request
        method:
          type: string
          description: HTTP method (defaults to GET)
        headers:
          type: object
          additionalProperties:
            type: string
          description: Request headers
        body:
          type: string
          description: Request body as text
        timeoutMs:
          type: integer
          description: Request timeout in milliseconds
        maxResponseBytes:
          type: integer
          description: Maximum response body size in bytes
        followRedirects:
          type: boolean
          description: Whether to follow HTTP redirects (defaults to true)
    HttpResponse:
      type: object
      required: [status, body]
      properties:
        status:
          type: integer
          description: HTTP status code
        headers:
          type: object
          additionalProperties:
            type: string
          description: Response headers
        body:
          type: string
          description: Response body as text
        url:
          type: string
          description: Final URL after any redirects
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        detail:
          type: string
          description: Additional context about the error
        code:
          type: string
          description: Machine-readable error code
security:
  - bearerAuth: []